Remove the counter code
The main purpose of the counter code is to count already saved datafiles so that a database could be organized for training a NN to detect features automatically. This feature should however go through some refactoring. Therefore, the counter code was moved to count.py under the aux directory for later.